Develop

+++ BlueBird/main/views.py
@@ -1,8 +1,12 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
 # Create your views here.
 
 #rooms = [
@@ -12,6 +16,29 @@ from .forms import RoomForm
 #    {'id': 4, 'name': 'Fourth room'}
 #]
 
+def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('/')
+
+    if request.method == 'POST':
+        username = request.POST.get("username")
+        password = request.POST.get("password")
+        try:
+            user = User.objects.get(username = username)
+        except:
+            messages.error(request, "User does not exist.")
+        user = authenticate(request, username = username, password = password)
+        if user is not None:
+            login(request, user)
+            return redirect('/')
+        else:
+            messages.error(request, "Username or password is incorrect.")
+    context = {}
+    return render(request, 'login.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('/')
 
 def index(request):
     #return HttpResponse("<h1> Response from views.py </h1>")
@@ -28,6 +55,7 @@ def room(request, pk):
     context = {'room': room_number}
     return render(request, 'room.html', context)
 
+@login_required(login_url = 'login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -39,9 +67,14 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'room_form.html', context)
 
+@login_required(login_url = 'login')
 def update_room(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to view this page.")
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
         if form.is_valid():
@@ -50,8 +83,13 @@ def update_room(request, pk):
     context = {'form': form}
     return render(request, 'room_form.html', context)
 
+@login_required(login_url = 'login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to view this page.")
+
     if request.method == 'POST':
         room.delete()
         return redirect('/')
